updated alert view, added safety json
