+++ src/app/components/views/weather-view/weather-view.component.html
@@ -8,57 +8,53 @@
       </div>
     </div>
   </div>
-</div>
-  <div class="alert-card">
+  <ng-container *ngIf="weatherAlertResponse?.allActiveWeatherAlerts?.length > 0">
     <ng-container *ngFor="let alert of weatherAlertResponse?.allActiveWeatherAlerts; let i = index">
-      <hr>
-      <div class="row mb-3 mt-4">
-        <div class="col">
-          <div class="row ms-1">
-            <div class="col">
-                <h4>Weather Alert</h4>
-              <div class="row">
-                {{alert?.headline}}
-              </div>
-            </div>
+      <div class="card my-4 border" [ngClass]="{'border-danger': alert?.event?.includes('Warning') || alert?.event?.includes('Extreme'),'border-warning': alert?.event?.includes('Watch')}">
+        <div class="card-header"
+          [ngClass]="{
+            'bg-danger': alert?.event?.includes('Warning') || alert?.event?.includes('Extreme'),
+            'text-light': alert?.event?.includes('Warning') || alert?.event?.includes('Extreme'),
+            'bg-warning': alert?.event?.includes('Watch')}">
+          <h4 class="card-title">{{alert?.event ? alert?.event : 'Weather Alert'}}</h4>
+        </div>
+        <div class="card-body">
+          <p class="card-text">{{alert?.headline}}</p>
+          <ng-container *ngIf="alert?.safetyInfo">
+            <a data-bs-toggle="modal" data-bs-target="#detailsModal" class="text-end text-decoration-none" (click)="setSafetyInfoForModal(alert?.safetyInfo)">Safe Work Practice</a>
+          </ng-container>
+          
+          <div class="collapse" [id]="'seeMore' + i">
+            <hr>
+            <p [ngStyle]="{'font-size': alert?.description?.length > 150 ? '0.9rem' : 'default'}">{{alert?.description}}</p>
           </div>
-          <div class="clean-hr my-1"></div>
-          <div class="row">
-            <div class="col collapse" [id]="'seeMore' + i">
-              <span class="font-smaller">{{alert?.description}}</span>
-              <div class="row">
-                <div class="col">
-                  <a class="text-decoration-none m-1">Safe Work Practice</a>
-                </div>
-              </div>
-            </div>
+          <hr>
+          <div class="text-center pt-2">
+            <a class="text-decoration-none text-dark" data-bs-toggle="collapse" [attr.data-bs-target]="'#seeMore' + i">Show Description</a>
           </div>
-            <div class="row">
-              <div class="col text-center">
-              <button class="btn btn-outline-secondary btn-sm m-1" type="button" data-bs-toggle="collapse" [attr.data-bs-target]="'#seeMore' + i" aria-expanded="false" aria-controls="seeMore">Show Description</button>
-              </div>
-            </div>
         </div>
       </div>
     </ng-container>
-  </div>
-<div class="modal fade" id="detailsModal" tabindex="-1" role="dialog">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h1 class="modal-title fs-5" id="exampleModalLabel">Weather Details</h1>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        <h6>Bad weather is bad to work in. Try working in good weather, because it's safer!</h6>
-        <br>
-        <h6>No job is so important and no service is so urgent that</h6>
-        <br>
-        <p>please transfer Mike to Port St Lucie</p>
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-        <button type="button" class="btn btn-primary" data-bs-dismiss="modal">Ok!</button>
+  </ng-container>
+  <div class="modal fade" id="detailsModal" tabindex="-1" role="dialog">
+    <div class="modal-dialog">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h1 class="modal-title fs-5">Working Safe</h1>
+          <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+        </div>
+        <div class="modal-body">
+            <p>{{modalSafetyInfo?.description}}</p>
+            <ul>
+              <ng-container *ngFor="let guideline of modalSafetyInfo?.guidelines">
+                <li>{{guideline}}</li>
+              </ng-container>
+            </ul>
+            <p>For More Info: <a class="text-decoration-none" [href]="modalSafetyInfo?.jobAidLink">{{modalSafetyInfo?.jobAidTitle}}</a> </p>
+        </div>
+        <div class="modal-footer">
+          <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Dismiss</button>
+        </div>
       </div>
     </div>
   </div>
