+++ src/app/safetyJobAids.json
@@ -3,7 +3,7 @@
       "description": "In some parts of our service area, snowstorms and/or ice storms can occur quickly, creating a number of potential hazards. Driving is particularly hazardous under these conditions, and one of the most dangerous situations is the potential for becoming stranded in an inaccessible area.",
       "guidelines": [
         "Be alert to National Weather Service Storm Watches and Warnings.",
-        "Dress appropriately for conditions. See EHS-3500-FST-7, Working in the Cold Fact Sheet.",
+        "Dress appropriately for conditions.",
         "Avoid isolated areas where you may be stranded by a storm.",
         "Be alert to changing weather conditions.",
         "Drive defensively, being especially aware of changing road conditions and inexperienced drivers.",
