+++ src/app/services/weather.service.ts
@@ -9,7 +9,7 @@ import { WeatherAlertResponse } from '../models/weather-alert.model';
   providedIn: 'root'
 })
 export class WeatherService {
-  private serverURL: string = 'https://api.weather.gov/alerts';
+  private serverURL: string = 'https://api.weather.gov/alerts/active';
   private apiResults: WeatherAlertResponse = null;
   private loadingChanged: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
   private loading: boolean;
